+++ formio/models/formio_builder.py
@@ -120,6 +120,15 @@ class Builder(models.Model):
         "Show User Metadata", tracking=True, help="Show submission and assigned user metadata in the Form header.", default=True)
     wizard = fields.Boolean("Wizard", tracking=True)
     wizard_on_next_page_save_draft = fields.Boolean("Wizard on Next Page Save Draft", tracking=True)
+    submission_url_add_url_query = fields.Selection(
+        string="Submission URL add URL query",
+        selection=[
+            ("window", "Window iframe (src)"),
+            ("window.parent", "Window parent (URL)"),
+        ],
+        tracking=True,
+        help="Enables adding the URL query string (params) from the window's iframe (src) or window.parent to the form submission URL endpoint.",
+    )
     translations = fields.One2many('formio.builder.translation', 'builder_id', string='Translations', copy=True)
     languages = fields.One2many('res.lang', compute='_compute_languages', string='Languages')
     allow_force_update_state_group_ids = fields.Many2many(
@@ -425,6 +434,7 @@ class Builder(models.Model):
             'portal_submit_done_url': self.portal_submit_done_url,
             'readOnly': self.is_locked,
             'wizard_on_next_page_save_draft': self.wizard and self.wizard_on_next_page_save_draft,
+            'submission_url_add_url_query': self.submission_url_add_url_query
         }
         return params
 
@@ -475,6 +485,7 @@ class Builder(models.Model):
         params = {
             'portal_submit_done_url': self.portal_submit_done_url,
             'wizard_on_next_page_save_draft': self.wizard and self.wizard_on_next_page_save_draft,
+            'submission_url_add_url_query': self.submission_url_add_url_query
         }
         return params
 
@@ -482,7 +493,8 @@ class Builder(models.Model):
         """ Form: Odoo JS (Owl component) misc. params """
         params = {
             'public_submit_done_url': self.public_submit_done_url,
-            'wizard_on_next_page_save_draft': self.wizard and self.wizard_on_next_page_save_draft
+            'wizard_on_next_page_save_draft': self.wizard and self.wizard_on_next_page_save_draft,
+            'submission_url_add_url_query': self.submission_url_add_url_query
         }
         return params
 
@@ -531,5 +543,8 @@ class Builder(models.Model):
                 i18n[code][trans.source] = trans.value
         return i18n
 
-    def _generate_odoo_domain(self, domain=[], data={}):
+    def _etl_odoo_data(self, params):
+        return {}
+
+    def _generate_odoo_domain(self, domain=[], params={}):
         return domain