+++ formio_data_api/models/formio_builder.py
@@ -3,6 +3,7 @@
 
 import logging
 
+from functools import reduce
 from formiodata.builder import Builder
 
 from odoo import fields, models, _
@@ -83,13 +84,13 @@ class FormioBuilder(models.Model):
                 display_error = _(msg) % (component.label, component.key, properties)
                 raise ValidationError(display_error)
 
-    def _get_formio_eval_context(self, component_server_api, formio_form=None, component=None, data={}):
+    def _get_formio_eval_context(self, component_server_api, formio_form=None, component=None, params={}):
         """ Prepare the context used when evaluating python code
 
         :param component_server_api: formio.component.server.api model record object
         :param formio_form: formio.form model record object
         :param component: formiodata Component object
-        :param data: possible dict with data, eg from URL query params
+        :param params: possible dict with data, eg from URL query params
             by the /data URL endpoint
         :returns: dict -- evaluation context given to safe_eval
         """
@@ -102,8 +103,54 @@ class FormioBuilder(models.Model):
             'dateutil': safe_eval.dateutil,
             'time': safe_eval.time,
             'component': component,
-            'data': data
+            'params': params
         }
         if component_server_api.type == 'values':
             res['values'] = {}
         return res
+
+    def _etl_odoo_data(self, params={}):
+        """
+        ETL (Odoo) data and prefill in the Form components.
+        """
+        res = super(FormioBuilder, self)._etl_odoo_data(params)
+        if bool(params):
+            res.update(self._etl_component_server_api(params=params))
+        return res
+
+    def _etl_component_server_api(self, params={}):
+        data = {}
+        api_values = {}
+        if self.component_server_api_ids.filtered('active'):
+            for comp_key, comp in self._formio.input_components.items():
+                prop_api = comp.properties.get('server_api')
+                prop_value = comp.properties.get('server_api_value')
+                prop_value_obj = comp.properties.get('server_api_value_obj')
+
+                if comp_key not in data and prop_api and prop_value:
+                    component_server_api = self.component_server_api_ids.filtered(lambda x: x.active and x.name == prop_api)
+                    if not component_server_api:
+                        _logger.error('NOT FOUND [formio.component.code.api] with name: %s' % prop_api)
+
+                    if component_server_api and api_values.get(component_server_api.name):
+                        value = api_values[component_server_api.name][prop_value]
+                        if prop_value_obj:
+                            # TODO-2: refactor DRY
+                            value_fields = prop_value_obj.split('.')
+                            value = reduce(getattr, value_fields, value)
+                        data[comp_key] = value
+                    elif component_server_api:
+                        eval_context = self._get_formio_eval_context(component_server_api, self, comp, params=params)
+                        # nocopy allows to return 'value'
+                        safe_eval.safe_eval(component_server_api.code, eval_context, mode="exec", nocopy=True)
+                        context_values = eval_context.get('values')
+                        value = context_values.get(prop_value)
+                        if prop_value_obj:
+                            # TODO-2: refactor DRY
+                            value_fields = prop_value_obj.split('.')
+                            value = reduce(getattr, value_fields, value)
+                        # caching
+                        if context_values:
+                            api_values[component_server_api.name] = context_values
+                        data[comp_key] = value
+        return data